Prevent and remove double-disposal in `DiscoveryService` and `RemoteConfigurationManager` (#7240)
## Summary of changes

- Removes instances of double-disposal
- Prevents double-disposal causing issues

## Reason for change

In https://github.com/DataDog/dd-trace-dotnet/pull/7235, we re-enabled
the automatic checking of integration test logs for errors which had
accidentally been broken during a refactor. When enabled, we discovered
multiple errors in the logs, e.g.

```log
Error cleaning up old tracer manager System.InvalidOperationException: An attempt was made to transition a task to a final state when it had already completed.
   at System.Threading.Tasks.TaskCompletionSource`1.SetResult(TResult result)
   at Datadog.Trace.Agent.DiscoveryService.DiscoveryService.DisposeAsync() in /_/tracer/src/Datadog.Trace/Agent/DiscoveryService/DiscoveryService.cs:line 315
   at Datadog.Trace.TracerManager.CleanUpOldTracerManager(TracerManager oldManager, TracerManager newManager) in /_/tracer/src/Datadog.Trace/TracerManager.cs:line 272
```

Tracing it through, there were 3 main sources of these logs

- https://github.com/DataDog/dd-trace-dotnet/pull/7200
- This removed exceptions from being thrown in the shutdown path by
replacing `CancellationTokenSource` with `TaskCompletionSource`.
However, a side-effect is that setting the result for a
`TaskCompletionSource` more than once throws an exception, while calling
`Cancel()` on a `CancellationTokenSource` is fine...
- https://github.com/DataDog/dd-trace-dotnet/pull/7220
- This made sure we _actually_ dispose tracer instances we create in our
integration tests properly, to avoid long-running background threads
from building up. However, in some cases we accidentally were disposing
the tracer _twice_.
- https://github.com/DataDog/dd-trace-dotnet/pull/3736
- This PR to `dd-trace` resulted in us disposing a `DiscoveryService`,
but then returning it for "normal" use, meaning it would be subsequently
disposed when the `Tracer` instance was disposed "normally".

## Implementation details

Fixed the three issues I found
- `StatsTests` and `TraceExporterTests` - don't have `using` _and_
`ShutdownAsync()`.
- Don't call `DiscoveryService.DisposeAsync` on the instance in
`dd-trace` when it's reused later
- Prevent double-disposal of `DiscoveryService` and
`RemoteConfigurationManager` from _actually_ causing issues

## Test coverage

I rebased https://github.com/DataDog/dd-trace-dotnet/pull/7235 on top of
this PR, but _Without_ the double-disposal protections, to ensure that
we have removed all existing cases, [and it's looking
good](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=182685&view=results).

Subsequently rebased on this full PR to confirm we're still looking good

## Other details

Ideally we would _always_ use `await using` but refactoring the
`StatsTests` was a pain, so I just kept the same structure (which is
effectively the same, as long as the stats tests don't throw). These
tests are very slow (20s+ each), so I may dig in and refactor them
subsequently.

The dd-trace `CiUtil` class is kind of a mess and very hard to follow
the logic, but I'll leave that up to @tonyredondo to look into in a
separate PR.

The other cases where we moved from CTS to TCS aren't problematic, as we
don't dispose them with the TracerManager, they're real singletons which
hook into the process exit directly. However, arguably, we should
consider making them double-disposal safe too.