Check for errors in the libdatadog log files (#7231)
## Summary of changes

- Ensures we check for any errors in the libdatadog files
- Add the `Datadog-Agent-State` header to responses

## Reason for change

To avoid errors creeping in unnoticed, we check for warnings and errors
in the log files from integration tests and smoke tests. Now that we are
starting to use libdatadog, we should check the libdatadog log files for
errors too.

The initial run identified that libdatadog expects the agent to send the
`Datadog-Agent-State` header, so added that to the mock agent.

## Implementation details

Added a libdatadog log parser and convert to the standard format

## Test coverage

Tested locally. This PR run is also a test. Will also do a full TFM and
installer test to see if anything got broken in the meantime 😬

## Other details

https://datadoghq.atlassian.net/browse/LANGPLAT-118

Discussed with libdatadog team whether the `Datadog-Agent-State` header
missing _should_ trigger an error, and they're going to discuss, but we
should probably add it here anyway. Additionally, we can use this
ourselves to reduce the impact of the discovery service