Add agent response handling to native interop (#7201)
## Summary of changes

Adds proper handling of the agent response using the libdatadog trace
exporter.

## Reason for change

We have observed memory leaks in testing when the data-pipeline tracer
exporter is used. Analysis by @hoolioh and @gleocadie identified one
potential source as being that memory for the agent response was being
allocated, but the managed side was not inspecting this and freeing the
memory.

Additionally, we noticed that updates to sampling rules, which are
returned by the agent, were not being passed back from libdatadog and
used to update the sampler.

## Implementation details

Update the implementation to call
`ddog_trace_exporter_response_get_body` and
`ddog_trace_exporter_response_free` to retrieve the body and free the
response.

The currently exposed `ddog_trace_exporter_response_get_body` is
sub-optimal, in that it doesn't provide the length of the body, so we
have to enumerate the data to find the length. That will be addressed
subsequently by adding a new API, once we have verified this resolves
the memory leak issue.

## Test coverage

https://github.com/DataDog/dd-trace-dotnet/pull/7165, which enables data
pipeline has been rebased on top of this PR to verify that this passes
all our existing tests, but it's possible that we don't have integration
tests that cover updating sampling rules...

If everything looks good after #7165 is rebased, we will merge this and
deploy to AAS reliability to evaluate if it solves the memory leak issue

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>