Add retry for build of unit tests on arm64 (#7259)
## Summary of changes

Retry _building_ the unit tests if it fails due to flake

## Reason for change

In #7207 we started retrying the _build_ stage of unit tests, because
the .NET SDK would sometimes randomly fail with a 139 exit code.
Unfortunately, I missed the `unit_tests_arm64` stage.

## Implementation details

Add a retry to the build stage only (not run)

## Test coverage

If this passes, we're ok