Add support for `xunit.v3` v3 (#7245)
## Summary of changes

Adds support for `xunit.v3` v3

## Reason for change

A new major version of `xunit.v3` was released

## Implementation details

Bumped all the numbers. 

## Test coverage

Updated tests to cover 3.x.
I dropped testing of 1.x to reduce the test load, but as that's the only
version of `xunit.v3` that supports .NET 6+ I don't know if we should or
not

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->