[Profiler/Tracer] Refactor the debug symbols folder setup (#7263)
## Summary of changes

Fix debug symbols publishing and refactor it.

## Reason for change
In a previous
[PR](https://github.com/DataDog/dd-trace-dotnet/pull/7246), we added
support for publishing PDB files. But it was added only for the manual
workflow.

This PR does it for when releasing a new version too.

## Implementation details

- Move debug symbols preparation in the action
- Clean up there and there

## Test coverage


https://github.com/DataDog/dd-trace-dotnet/actions/runs/16466188593/job/46544059599

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->