Use `TrySetResult` instead of `SetResult` (#7265)
## Summary of changes

Use `TaskCompletionSource.TrySetResult` instead of `SetResult` (which
throws)

## Reason for change

In #7200 we replaced `CancellationTokenSource` with
`TaskCompletionSource`, but this caused some exceptions on shutdown due
to double-disposal. In #7240 we fixed that, but did it in a way that
doesn't make sense. We don't need to use `Interlocked.Exchange` we can
just use `TrySetResult`.

## Implementation details

- Use `TrySetResult` instead of `SetResult` as the latter just calls the
former and throws
- Did a sweep to update some other usages of `SetResult` which are
unnecessarily risky

## Test coverage

Covered by existing test

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->