More fixes for do-not-merge (#7274)
## Summary of changes

Try to make the do-not-merge action a required job

## Reason for change

For some mystical reason, the do_not_merge job runs on PRs, but I can't
mark it as required. AI thinks this change will work. We'll see...

## Implementation details

Move the decision on pass/fail to a separate job

## Test coverage

This is the test, but I might have to merge it to be sure