Fix potential deadlocks in `IisFixture` process code (#7281)
## Summary of changes

Fixes potential deadlocks in the IIS fixture code 

## Reason for change

While working on the .NET branch, we had to update xunit. We
subsequently started seeing deadlocks in the Iis tests. I noticed
potential issues with the code that reads the process output (because I
fixed something similar in #7116 recently) and voila, no deadlocks 🎉

## Implementation details

Use the event-based handling already available in `ProcessHelper`
instead of spawning separate dedicated threads for handling reading the
output. Also some minor additional things:

- Allowed specifying a time to wait after calling `Process.Kill()` in
`ProcessHelper` before returning (we need this for the IIS Fixture)
- ~Make sure we dispose the `Process` object held by `ProcessHelper`~
This _is_ already disposed, because it gets "owned" by the
`ProcessResult` object instead

## Test coverage

Covered by existing tests - if it passes, we're good

## Other details

Following are dependent on this PR:
- #7170 
- #7280

https://datadoghq.atlassian.net/browse/LANGPLAT-632