Swap disposal ordering of inferred span (#7293)
## Summary of changes

Swaps the ordering when disposing the API Gateway inferred span to be
after the "root" span.
Realistically the API Gateway span will be the parent of the root span
for the ASP NET Core instrumentation.

## Reason for change

If the proxy scope is the parent of the `ASPNET Core` scope AND that
`ASPNET Core` scope is active this was hitting the following [piece of
code](https://github.com/DataDog/dd-trace-dotnet/blob/1752abdc32c02c702b7c8374bfd1bae8b381d7db/tracer/src/Datadog.Trace/AsyncLocalScopeManager.cs#L44-L48)
in `Async

```csharp
            if (current == null || current != scope)
            {
                // This is not the current scope for this context, bail out
                return;
            }
```

This would cause the scope for the proxy to remain open 😢 

## Implementation details

Swapped the ordering of them.

## Test coverage

- Identified and covered by existing tests and was caught in
[this](https://github.com/DataDog/dd-trace-dotnet/pull/7280) PR when
updating xUnit

## Other details
<!-- Fixes #{issue} -->

Cherry picked out and making it's own PR instead of bundling into the
[xUnit update one
](https://github.com/DataDog/dd-trace-dotnet/pull/7284) because it
passes and works without that.

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->