[Test Optimization] Fix base branch SHA usage (#7300)
## Summary of changes

This PR fixes the use of the base branch sha value, we thought this sha
was the BASE sha, but we realized it is the HEAD sha of the base branch.

## Reason for change

Jira: https://datadoghq.atlassian.net/browse/SDTEST-2352

## Implementation details

Small updates on the tag names

## Test coverage

Update the tests json files and small modification on the github test.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->