Filter connection header in flaky tests (#7290)
## Summary of changes

Some security WebApi tests that rely on the Sample.security.WebAPI (.NET
framework) does not show consistent result. The result of the test
depends on the order of the tests. The problem is that in some requests,
the Connection header is returned (or not) by the framework depending on
the request order. That causes the generated security header fingerprint
to vary.

The fingerprint hash and format is defined
[here](https://docs.google.com/document/d/1DivOa9XsCggmZVzMI57vyxH2_EBJ0-qqIkRHm_sEvSs/edit?tab=t.0).

In order to avoid that, a scrubber has been added to skip that
particular flag in the fingerprint.

Other approaches have been tried to get a consistent result that failed:
* Adding a filter in FilterConfig.cs in the sample to overwrite the
connection header.
* Set UseProxy = false in HttpClientHandler in the test
* Moved to an HttpModule and added Connection: keep-alive in EndRequest

In all these cases, the framework seemed to overwrite the connection
header value even when set explicitly.

## Reason for change

Avoid having tests that depend on test execution order.

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->