Add `[MethodImpl(MethodImplOptions.NoInlining]` to all our instrumented APIs in Datadog.Trace.Manual (#7302)
## Summary of changes

Adds `[MethodImpl(MethodImplOptions.NoInlining]` to avoid race
conditions on app startup

## Reason for change

There’s essentially a race condition with how the .NET runtime profiling
APIs work and how we inject into a customer’s application. We typically
inject into the start of the `Main()` method, and then request ReJIT for
all the methods that we want to instrument. The next time a target
method is called, the runtime will see the ReJIT request, and add our
instrumentation.

Another aspect is inlining; if a method that we wanted to instrument has
been inlined into the caller, then just doing a ReJIT won't work
(because the method is "gone" in this case). We work around that for
normally by blocking inlining for methods that we’re interested in (i.e.
methods we instrument). However, if the method we’re interested in has
_already_ been inlined _before_ we are injected then we could miss calls
at the start of the app.

Essentially there's a race condition between:
- Calls in `Main()` being JITed (and inlined).
- Our startup hook being injected and processing the instrumentation
list to block inlining for methods we care about.

We currently can't do much about this for arbitrary target methods (we
have some ideas, but that's a separate issue!) but for our manual APIs
we know we _will_ want to mark them as not-inlined. So we can
essentially pre-emptively mark these APIs with
`[MethodImpl(MethodImplOptions.NoInlining]`, exculding these methods
from the race condition (theoretically; it's technically up to the JIT
if it chooses to honour the attributes IIRC)

## Implementation details

- Add `[MethodImpl(MethodImplOptions.NoInlining]` to all the methods
marked with `[Instrumented]`
- This is a bit ugly for properties because they need to be applied to
the getter/setter, not the property
  - For consistency, moved the `[Instrumented]` attribute there too

## Test coverage

Updated the unit test to ensure that every instrumented method is
explicitly marked as `NoInlining`. Confirmed that this catches issues if
you miss one

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->