[IAST] Remove troublesome string.Concat overloads with ReadOnlySpan<char> (#7318)
## Summary of changes
The method of obtaining the underlying string from a
`ReadOnlySpan<char>` could crash if it came from a stack allocated char
array, and there is no reliable way of telling them apart, so it is
safer to remove the `String.Concat` overloads that used it.

## Reason for change
Some crashes were detected

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->