Fix resource-based sampling for ASP.NET Core (#7316)
## Summary of changes

Eagerly calculate a stand-in resource name if resource-based sampling is
being used

## Reason for change

In https://github.com/DataDog/dd-trace-dotnet/pull/6936, we fixed an
issue where resource-based sampling wasn't working in IIS. This was
because we can't calculate the "final" resource name for a span until
after the MVC framework has run. Previously, we were delaying the
resource name calculation to this point. For the fix, we calculated a
provisional resource name, which we subsequently change. In #7312 we
updated the logic to only generate the resource name when resource-based
sampling was being used.

This PR fixes the same issue for ASP.NET Core

## Implementation details

- Added unit and integration tests to demonstrate the issue
- For the integration tests, we have to make a sampling decision after
the pipeline starts, but before routing is hit.
- To avoid creating new samples or breaking existing ones, I manually
forced a sampling decision to the existing `PingMiddleware` (which is
invoked before routing kicks in).
- The initial commit contains the snapshots in their "incorrect" state.
The second commit shows the "diff" showing the sampling working
- Make resource name generation eager in `AspNetCoreDiagnosticObserver`
if we have resource-based sampling

## Test coverage

Unit tests + integration tests, IIS vs standalone, In and out of process

## Other details

Part of a stack

- https://github.com/DataDog/dd-trace-dotnet/pull/7311
- https://github.com/DataDog/dd-trace-dotnet/pull/7312
- https://github.com/DataDog/dd-trace-dotnet/pull/7316 👈

https://datadoghq.atlassian.net/browse/APMS-16673