[Test Optimization] Avoid crashing on Allure.Xunit usage (#7305)
## Summary of changes

This PR avoid crashing testhost when Allure.XUnit is being used.

## Reason for change

Currently using Allure in your XUnit tests will fail.

## Implementation details

Ensure to bypass some messages to the messagebus in order to trigger
some XUnit internal events in the MessageSink. Also when we disable a
test we stop sending messages in the messagebus regarding that test.

## Test coverage

Added the Allure.XUnit package in one of the XUnit sample test. Also it
tries to add the allure api inside a test. The goal is to don't crash
the samples app and have the same test passing without any exception
(errors in the logs).

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->