[AAP] Add resource name to the collected endpoints (#7317)
## Summary of changes

Added the ResourceName to the endpoints collected based on the last
update of the
[RFC](https://docs.google.com/document/d/1txwuurIiSUWjYX7Xa0let7e49XKW2uhm1djgqjl_gL0).
## Reason for change

## Implementation details

- The resource name is defined as `method + path` except in the case of
an endpoint detected as a wildcard, the resource name is set to the
path.

## Test coverage

Updated the tests.

## Other details

#6733 

<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->