Refactor the profiling configuration reading (#7303)
## Summary of changes

Refactors how we determine the profiler's current configuration state

## Reason for change

Today, the profiler determines if it's enabled or disabled independently
of the tracer, in the native code. We then try to mirror that state in
the managed code, for several reasons

- To know whether we need to send app-domain information to the profiler
(e.g. service name, env)
- To know whether we need to send thread-context-change events to the
profiler (for CodeHotspots/EndpointProfiling)
- For telemetry reasons
- For recording the state in DSM payloads

As part of the hands-off config work, we will calculate whether to
enable the profiler based on the _full_ managed config, instead of just
using environment variables. However, when managed activation is
disabled (for back-compat investigation purposes), we _do_ need to just
use env vars, to "mirror" the state that will be calculated by the
native continuous profiler, essentially working in the existing "legacy"
state.

The refactoring in this PR is intended to lay the groundwork for the
future hands-off config work in the profiler, as well as to solve some
existing issues with inconsistent/missing telemetry

## Implementation details

- Add a `ProfilerState` `enum` to track the profiler as a 3-state
enablement which represents its real state, instead of trying to force
it as a boolean
- Use `EnvironmentConfigurationSource` to record variables read from env
in telemetry correctly.
- Add support for the `DD_PROFILER_MANAGED_ACTIVATION_ENABLED` which
will be used shortly

## Test coverage

Added unit tests for the new `ProfilerSettings` which encapsulates most
of the logic

## Other details

This PR technically requires support for
`DD_PROFILER_MANAGED_ACTIVATION_ENABLED` has been merged in order to
report the correct values. This isn't a big deal, as it will likely be
merged soon, but if it isn't, we should tweak the logic in
`ProfilerSettings` so that it's correct. Alternatively, I can split this
into two PRs - one that is correct _today_, and one that is correct only
with the profiler PR

---------

Co-authored-by: chrisnas <chrisnas@users.noreply.github.com>