Update xunit to latest, fix analyzer warnings, and update analyzer tests (#7280)
## Summary of changes

- Update `xunit` to latest v2 package version (`v2.9.3`)
- Update Analyzer testing libraries to the new test-framework agnostic
versions
- Fix all the errors xunit now flags

## Reason for change

As part of the .NET 10 work we tried updating to the latest xunit. It's
not _entirely_ clear if that work was actually necessary (there appeared
to be a bug in preview 5 that didn't impact preview 6) but rather than
waste the effort, pulling this out into a separate PR.

## Implementation details

Mostly just bumping package versions. For the analyzer library changes,
I followed [this
guide](https://github.com/dotnet/roslyn-sdk/blob/main/src/Microsoft.CodeAnalysis.Testing/README.md)
to update. For xunit I bumped required packages, and then fixed a bunch
of errors the xunit.analyzers started flagging:

- `Assert.Empty` -> use FluentAssertions
- "Incorrect type" -> Use `int` instead of `HttpStatusCode`,
`NullableBooleanTestCases` instead of `BooleanTestCases` (for example)
- "Don't use `ConfigureAwait(false)`" -> Remove it (xunit has a sync
context)
- "xUnit1031 - Test methods should not use Blocking operations" - I
_mostly_ ignored this one, because the tests are in testing for
low-level stuff (call target machinery) and didn't want to change the
behavior

## Test coverage

This is the test. Need to double check the tests are definitely
_actually_ running though...

## Other details

Related to
- #7170 
- #7281
- https://datadoghq.atlassian.net/browse/LANGPLAT-632