Work around flake in the Couchbase 3 integration tests (#7357)
## Summary of changes

Work around flake in the Couchbase 3 integration tests

## Reason for change

In #7029 we added code to workaround infra flake in the couchbase tests
by looking for `AuthenticationException`. However, in Couchbase 3,
`AuthenticationFailureException` is thrown instead, so we should look
for that too.

Additionally, I noticed that if I just run the sample locally, with no
couchbase running, a totally different issue arises. Working abound that
by trying to explicitly connect, and timing out if that fails

## Implementation details

Add `AuthenticationFailureException` and `UnambiguousTimeoutException`
to the ignored list. Re-throwing (instead of swallowing) when an error
occurs during auth (as we can't continue)

## Test coverage

Tested locally, and confirmed we return the skip code.