Cache response in order to avoid unnecessary deserializations (#7320)
## Summary of changes
Cache agent's response to avoid unnecessary deserializations in case
sample rates haven't changed from previous response.
## Reason for change
Reduce performance overhead.
## Implementation details

## Test coverage

## Other details
This PR implements part of the functionality of this
[one](https://github.com/DataDog/dd-trace-dotnet/pull/7251). As
commented the rates versioning payload will be handled internally in the
native exporter but until it is available the response will be cached in
the same way as in the managed exporter.
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->