Try to fix flake by using unique keys in StackExchange sample tests (#7359)
## Summary of changes

Try to fix flake by using unique keys in StackExchange sample tests

## Reason for change

Saw some flake in the stackexchange tests, with a diff looking like
this:

```diff
   {
     TraceId: Id_107,
     SpanId: Id_108,
     Name: redis.command,
     Resource: HINCRBYFLOAT,
     Service: Samples.StackExchange.Redis,
     Type: redis,
+     Error: 1,
     Tags: {
       component: StackExchangeRedis,
       env: integration_tests,
+       error.msg: ERR hash value is not a float,
+       error.stack:
+ StackExchange.Redis.RedisServerException: ERR hash value is not a float
+ at StackExchange.Redis.ConnectionMultiplexer.ExecuteSyncImpl[T](Message message, ResultProcessor`1 processor, ServerEndPoint server, T defaultValue),
+       error.type: StackExchange.Redis.RedisServerException,
       language: dotnet,
       out.host: stackexchangeredis,
       out.port: 6379,
       peer.service: stackexchangeredis,
       redis.raw_command: HINCRBYFLOAT StackExchange.Redis.Database.Hash,
       runtime-id: Guid_1,
       span.kind: client,
       version: 1.0.0,
       _dd.peer.service.source: out.host
     },
     Metrics: {
       db.redis.database_index: 1.0,
       process_id: 0,
       _dd.top_level: 1.0,
       _dd.tracer_kr: 1.0,
       _sampling_priority_v1: 1.0
     }
   },
```

I _think_ the problem is that we are performing multiple operations
against the same hash in the "sync" tests, and some ordering of commands
it didn't like

## Implementation details

Use a unique command for each test, like we do in the async version

## Test coverage

Covered by existing

## Other details

Originally failed
[here](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=185286&view=logs&j=a69d05c5-a1c4-5b6f-4792-656fb54d97d4&t=7f0459b0-1f69-5432-c3d8-9b1decfa9496)


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->